Moved to geoptz wrapper
