+++ src/modules/visualization/sidebar/components/PanelVisualizationWrapper.vue
