+++ src/modules/visualization/sidebar/composables/useVisualizationCleanup.ts
