+++ src/modules/visualization/visualizations/chart/Chart.vue
@@ -11,7 +11,7 @@ import {
 	ICurveLayerProperties,
 	ISweApiDataSourceProperties,
 } from '@/lib/VisualizationHelpers';
-import { useVisualizationCleanup } from '../../components/composables/useVisualizationCleanup';
+import { useVisualizationCleanup } from '../../sidebar/composables/useVisualizationCleanup';
 
 const props = defineProps<{
 	visualization: OSHVisualization;
