+++ src/modules/visualization/visualizations/text/Text.vue
@@ -5,7 +5,7 @@ import { randomUUID } from 'osh-js/source/core/utils/Utils.js';
 import { DATASOURCE_DATA_TOPIC } from 'osh-js/source/core/Constants.js';
 import { createDatasource } from '@/modules/visualization/services/datasource.service';
 import { ISweApiDataSourceProperties } from '@/lib/VisualizationHelpers';
-import { useVisualizationCleanup } from '../../components/composables/useVisualizationCleanup';
+import { useVisualizationCleanup } from '../../sidebar/composables/useVisualizationCleanup';
 
 // Generate a random ID when the component is created
 const textboxId = ref('textbox-' + randomUUID());
